backend/pyproject.toml
[project]
name = "the_likes_backend"
version = "0.1"
description = "Backend for The Likes app. Find people with same values."
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "aiosmtplib>=4.0.1",
    "alembic>=1.16.4",
    "async-lru>=2.0.5",
    "asyncpg>=0.30.0",
    "celery[redis]>=5.5.3",
    "click==8.2.1",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "geoalchemy2[shapely]>=0.18.0",
    "geopy>=2.4.1",
    "gunicorn>=23.0.0",
    "loguru>=0.7.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pwnedpasswords>=2.0.0",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-multipart>=0.0.20",
    "redis[hiredis]>=5.2.1",
    "sqlalchemy>=2.0.41",
    "typer>=0.16.0",
    "uvicorn[standard]>=0.35.0",
    "websockets>=15.0.1",
    "zxcvbn-python>=4.4.24",
]

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "jupyter>=1.1.1",
    "pytest-asyncio>=1.2.0",
    "pytest-order>=1.3.0",
    "random-password-generator>=2.2.0",
    "ruff>=0.12.7",
]

[tool.ruff]
line-length = 79
lint.unfixable = ["F401"]  # unused imports
lint.select = ["E4", "E7", "E9", "F"]

[tool.ruff.format]
quote-style = "single"
